refactor(webui): split backup/restore and OTA out of webui.cpp
webui.cpp held every HTTP handler in the firmware across 3700 lines, which
made it the file nobody wants to open. The two largest groups are also the
two most self-contained, so they move out first:

  webui.cpp         3700 -> 1977 lines
  webui_backup.cpp  1174 lines  (/api/backup, /api/restore)
  webui_ota.cpp      692 lines  (/ota_update, /api/ota_status,
                                 /api/restart, /api/factory-reset)

Restart and factory reset go with the OTA unit rather than a "system" one:
all four take the exclusive device-operation reservation and all end in a
controlled reboot, and separating them would put prepare_ota_heap() in a
different file from its only callers.

webui_internal.h is the seam. It exposes only what the extracted units
consume - the request helpers, the settings serializer, the two context
accessors and ScopedOperationReservation - so the module's internals do not
become public API. The context stays static in webui.cpp behind
webui_settings() / webui_status_led(), which keeps that file the only writer.
ScopedOperationReservation moved into the header and gained deleted copy and
assignment operators; it was never copied, and now it cannot be.

This commit is a pure move, kept reviewable as one: the OTA block is
byte-identical to its previous form, and the backup block differs only by
the removal of the relocated class. The registered route set and the
handler-function set are identical to the previous commit, checked by
extracting and diffing both from the old and new trees. clang-format is
applied separately in the next commit rather than mixed in here, because
reformatting 1700 moved lines in the same diff would make exactly this
verification impossible for a reviewer.

Firmware size grows 672 bytes, from 0x131890 to 0x131b30: the same code in
three translation units gives the compiler fewer inlining opportunities.

Note for anyone repeating this: main/CMakeLists.txt globs its sources, so a
new .cpp needs `idf.py reconfigure` before it is compiled at all. Without it
the build fails at link time with undefined references to the new file's
symbols, which reads like a linkage bug rather than a stale glob.
