+++ main/webui.cpp
@@ -32,6 +32,7 @@
 #include <new>
 #include <stdarg.h>
 #include "webui.h"
+#include "webui_internal.h"
 #include "ping_service.h"
 #include "esp_log.h"
 #include "esp_system.h"
@@ -69,11 +70,6 @@ static const char *TAG = "WebUI";
 // Safe cJSON number accessor: returns the integer value only when the item
 // exists and is actually a number. Prevents undefined behaviour when the
 // frontend sends an unexpected type.
-static inline int cJSON_GetIntValueSafe(cJSON *item, int defaultValue)
-{
-    return (item && cJSON_IsNumber(item)) ? item->valueint : defaultValue;
-}
-
 #define EMBED_HANDLER(_uri, _resource, _contentType, _contentEncoding) \
     extern const char _resource[] asm("_binary_" #_resource "_start"); \
     extern const size_t _resource##_length asm(#_resource "_length");  \
@@ -111,6 +107,12 @@ static RawUartUdpListener *_rawUartUdpListener;
 static RadioModuleConnector *_radioModuleConnector;
 static RadioModuleDetector *_radioModuleDetector;
 static char _token[46];
+
+// Read-only handles for the handler units extracted out of this file
+// (webui_backup.cpp, webui_ota.cpp). They stay static here so this
+// translation unit remains the only writer; see webui_internal.h.
+Settings *webui_settings() { return _settings; }
+LED *webui_status_led() { return _statusLED; }
 
 static void emit_log_enable_snapshot();
 int recv_full_body(httpd_req_t *req, char *buf, size_t buf_size);
@@ -245,7 +247,7 @@ static esp_err_t generate_fresh_admin_token(char *out, size_t out_size)
     return ESP_OK;
 }
 
-static esp_err_t rotate_admin_token()
+esp_err_t rotate_admin_token()
 {
     if (!_settings) return ESP_ERR_INVALID_STATE;
     char fresh_token[sizeof(_token)] = {};
@@ -933,8 +935,8 @@ bool cJSON_GetBoolValue(const cJSON *item)
     return (item && cJSON_IsBool(item)) ? cJSON_IsTrue(item) : false;
 }
 
-static esp_err_t send_json_error(httpd_req_t *req, const char *status,
-                                 const char *code, const char *field = NULL)
+esp_err_t send_json_error(httpd_req_t *req, const char *status,
+                          const char *code, const char *field)
 {
     cJSON *root = cJSON_CreateObject();
     if (!root) {
@@ -957,7 +959,7 @@ static esp_err_t send_json_error(httpd_req_t *req, const char *status,
     return result;
 }
 
-static bool valid_admin_username(const char *value)
+bool valid_admin_username(const char *value)
 {
     if (!value) return false;
     const size_t length = strlen(value);
@@ -972,8 +974,8 @@ static bool valid_admin_username(const char *value)
     return true;
 }
 
-static bool parse_ipv4_json(const cJSON *item, bool optional,
-                            ip4_addr_t *address)
+bool parse_ipv4_json(const cJSON *item, bool optional,
+                     ip4_addr_t *address)
 {
     if (!item || !cJSON_IsString(item) || !address) return false;
     const char *value = cJSON_GetStringValue(item);
@@ -984,7 +986,7 @@ static bool parse_ipv4_json(const cJSON *item, bool optional,
     return value && ip4addr_aton(value, address) != 0;
 }
 
-static bool refresh_restart_sync_from_settings()
+bool refresh_restart_sync_from_settings()
 {
     const bool enabled = _settings && _settings->getFlashPause();
     set_flash_pause_enabled(enabled);
@@ -1421,1755 +1423,6 @@ httpd_uri_t post_settings_json_handler = {
     .method = HTTP_POST,
     .handler = post_settings_json_handler_func,
     .user_ctx = NULL};
-
-static bool backup_get_bool(const cJSON *object, const char *key, bool *value)
-{
-    const cJSON *item = cJSON_GetObjectItemCaseSensitive(object, key);
-    if (!cJSON_IsBool(item) || !value) return false;
-    *value = cJSON_IsTrue(item);
-    return true;
-}
-
-static bool backup_get_uint(const cJSON *object, const char *key,
-                            unsigned maximum, unsigned *value)
-{
-    const cJSON *item = cJSON_GetObjectItemCaseSensitive(object, key);
-    if (!cJSON_IsNumber(item) || !value ||
-        item->valuedouble < 0 || item->valuedouble > maximum ||
-        item->valuedouble != item->valueint) {
-        return false;
-    }
-    *value = static_cast<unsigned>(item->valueint);
-    return true;
-}
-
-static bool backup_get_string(const cJSON *object, const char *key,
-                              char *value, size_t value_size)
-{
-    const cJSON *item = cJSON_GetObjectItemCaseSensitive(object, key);
-    if (!cJSON_IsString(item) || !item->valuestring ||
-        !value || value_size == 0 ||
-        strlen(item->valuestring) >= value_size) {
-        return false;
-    }
-    snprintf(value, value_size, "%s", item->valuestring);
-    return true;
-}
-
-static esp_err_t add_monitoring_backup(cJSON *root)
-{
-    monitoring_config_t *config =
-        new (std::nothrow) monitoring_config_t();
-    if (!config) return ESP_ERR_NO_MEM;
-    if (monitoring_get_config(config) != ESP_OK) {
-        delete config;
-        return ESP_ERR_INVALID_STATE;
-    }
-
-    cJSON *monitoring = cJSON_AddObjectToObject(root, "monitoring");
-    cJSON *checkmk = monitoring
-        ? cJSON_AddObjectToObject(monitoring, "checkmk") : NULL;
-    cJSON *mqtt = monitoring
-        ? cJSON_AddObjectToObject(monitoring, "mqtt") : NULL;
-    cJSON *prometheus = monitoring
-        ? cJSON_AddObjectToObject(monitoring, "prometheus") : NULL;
-    cJSON *syslog = monitoring
-        ? cJSON_AddObjectToObject(monitoring, "syslog") : NULL;
-    cJSON *notify = monitoring
-        ? cJSON_AddObjectToObject(monitoring, "notify") : NULL;
-    if (!monitoring || !checkmk || !mqtt || !prometheus ||
-        !syslog || !notify) {
-        delete config;
-        return ESP_ERR_NO_MEM;
-    }
-
-    cJSON_AddBoolToObject(checkmk, "enabled", config->checkmk.enabled);
-    cJSON_AddNumberToObject(checkmk, "port", config->checkmk.port);
-    cJSON_AddStringToObject(checkmk, "allowedHosts",
-                            config->checkmk.allowed_hosts);
-
-    cJSON_AddBoolToObject(mqtt, "enabled", config->mqtt.enabled);
-    cJSON_AddStringToObject(mqtt, "server", config->mqtt.server);
-    cJSON_AddNumberToObject(mqtt, "port", config->mqtt.port);
-    cJSON_AddStringToObject(mqtt, "user", config->mqtt.user);
-    cJSON_AddStringToObject(mqtt, "password", config->mqtt.password);
-    cJSON_AddStringToObject(mqtt, "topicPrefix", config->mqtt.topic_prefix);
-    cJSON_AddBoolToObject(mqtt, "haDiscoveryEnabled",
-                          config->mqtt.ha_discovery_enabled);
-    cJSON_AddStringToObject(mqtt, "haDiscoveryPrefix",
-                            config->mqtt.ha_discovery_prefix);
-    cJSON_AddBoolToObject(mqtt, "tlsEnable", config->mqtt.tls_enable);
-    cJSON_AddBoolToObject(mqtt, "tlsSkipVerify",
-                          config->mqtt.tls_skip_verify);
-    cJSON_AddStringToObject(mqtt, "tlsCaCerts",
-                            config->mqtt.tls_ca_certs);
-    cJSON_AddStringToObject(mqtt, "tlsCertfile",
-                            config->mqtt.tls_certfile);
-    cJSON_AddStringToObject(mqtt, "tlsKeyfile",
-                            config->mqtt.tls_keyfile);
-    cJSON_AddBoolToObject(mqtt, "commandEnabled",
-                          config->mqtt.command_enabled);
-    cJSON_AddStringToObject(mqtt, "commandToken",
-                            config->mqtt.command_token);
-
-    cJSON_AddBoolToObject(prometheus, "enabled",
-                          config->prometheus.enabled);
-    cJSON_AddNumberToObject(prometheus, "port", config->prometheus.port);
-    cJSON_AddStringToObject(prometheus, "allowedHosts",
-                            config->prometheus.allowed_hosts);
-
-    cJSON_AddBoolToObject(syslog, "enabled", config->syslog.enabled);
-    cJSON_AddStringToObject(syslog, "server", config->syslog.server);
-    cJSON_AddNumberToObject(syslog, "port", config->syslog.port);
-    cJSON_AddNumberToObject(syslog, "transport",
-                            config->syslog.transport);
-    cJSON_AddNumberToObject(syslog, "minSeverity",
-                            config->syslog.min_severity);
-    cJSON_AddStringToObject(syslog, "hostname",
-                            config->syslog.hostname);
-
-    cJSON_AddBoolToObject(notify, "enabled", config->notify.enabled);
-    cJSON_AddNumberToObject(notify, "channels", config->notify.channels);
-    cJSON_AddStringToObject(notify, "webhookUrl",
-                            config->notify.webhook_url);
-    cJSON_AddStringToObject(notify, "webhookSecret",
-                            config->notify.webhook_secret);
-    cJSON_AddStringToObject(notify, "telegramToken",
-                            config->notify.telegram_token);
-    cJSON_AddStringToObject(notify, "telegramChatId",
-                            config->notify.telegram_chatid);
-    cJSON_AddStringToObject(notify, "smtpServer",
-                            config->notify.smtp_server);
-    cJSON_AddNumberToObject(notify, "smtpPort",
-                            config->notify.smtp_port);
-    cJSON_AddNumberToObject(notify, "smtpTls",
-                            config->notify.smtp_tls);
-    cJSON_AddStringToObject(notify, "smtpUser",
-                            config->notify.smtp_user);
-    cJSON_AddStringToObject(notify, "smtpPassword",
-                            config->notify.smtp_password);
-    cJSON_AddStringToObject(notify, "smtpFrom",
-                            config->notify.smtp_from);
-    cJSON_AddStringToObject(notify, "smtpTo",
-                            config->notify.smtp_to);
-    cJSON_AddNumberToObject(notify, "cooldownSeconds",
-                            config->notify.cooldown_seconds);
-    cJSON_AddNumberToObject(notify, "eventMask", config->notify.event_mask);
-
-    delete config;
-    return ESP_OK;
-}
-
-static esp_err_t parse_monitoring_backup(
-    const cJSON *root, monitoring_config_t **restored_config)
-{
-    if (!restored_config) return ESP_ERR_INVALID_ARG;
-    *restored_config = NULL;
-
-    const cJSON *monitoring =
-        cJSON_GetObjectItemCaseSensitive(root, "monitoring");
-    if (!monitoring) return ESP_OK; // Backwards-compatible legacy backup.
-    if (!cJSON_IsObject(monitoring)) return ESP_ERR_INVALID_ARG;
-
-    const cJSON *checkmk =
-        cJSON_GetObjectItemCaseSensitive(monitoring, "checkmk");
-    const cJSON *mqtt =
-        cJSON_GetObjectItemCaseSensitive(monitoring, "mqtt");
-    const cJSON *prometheus =
-        cJSON_GetObjectItemCaseSensitive(monitoring, "prometheus");
-    const cJSON *syslog =
-        cJSON_GetObjectItemCaseSensitive(monitoring, "syslog");
-    const cJSON *notify =
-        cJSON_GetObjectItemCaseSensitive(monitoring, "notify");
-    if (!cJSON_IsObject(checkmk) || !cJSON_IsObject(mqtt) ||
-        !cJSON_IsObject(prometheus) || !cJSON_IsObject(syslog) ||
-        !cJSON_IsObject(notify)) {
-        return ESP_ERR_INVALID_ARG;
-    }
-
-    monitoring_config_t *config =
-        new (std::nothrow) monitoring_config_t();
-    if (!config) return ESP_ERR_NO_MEM;
-    memset(config, 0, sizeof(*config));
-    unsigned number = 0;
-
-    bool valid =
-        backup_get_bool(checkmk, "enabled", &config->checkmk.enabled) &&
-        backup_get_uint(checkmk, "port", UINT16_MAX, &number);
-    if (valid) config->checkmk.port = static_cast<uint16_t>(number);
-    valid = valid && validatePort(config->checkmk.port) &&
-        backup_get_string(checkmk, "allowedHosts",
-                          config->checkmk.allowed_hosts,
-                          sizeof(config->checkmk.allowed_hosts));
-
-    valid = valid &&
-        backup_get_bool(mqtt, "enabled", &config->mqtt.enabled) &&
-        backup_get_string(mqtt, "server", config->mqtt.server,
-                          sizeof(config->mqtt.server)) &&
-        backup_get_uint(mqtt, "port", UINT16_MAX, &number);
-    if (valid) config->mqtt.port = static_cast<uint16_t>(number);
-    valid = valid && validatePort(config->mqtt.port) &&
-        backup_get_string(mqtt, "user", config->mqtt.user,
-                          sizeof(config->mqtt.user)) &&
-        backup_get_string(mqtt, "password", config->mqtt.password,
-                          sizeof(config->mqtt.password)) &&
-        backup_get_string(mqtt, "topicPrefix", config->mqtt.topic_prefix,
-                          sizeof(config->mqtt.topic_prefix)) &&
-        backup_get_bool(mqtt, "haDiscoveryEnabled",
-                        &config->mqtt.ha_discovery_enabled) &&
-        backup_get_string(mqtt, "haDiscoveryPrefix",
-                          config->mqtt.ha_discovery_prefix,
-                          sizeof(config->mqtt.ha_discovery_prefix)) &&
-        backup_get_bool(mqtt, "tlsEnable", &config->mqtt.tls_enable) &&
-        backup_get_bool(mqtt, "tlsSkipVerify",
-                        &config->mqtt.tls_skip_verify) &&
-        backup_get_string(mqtt, "tlsCaCerts",
-                          config->mqtt.tls_ca_certs,
-                          sizeof(config->mqtt.tls_ca_certs)) &&
-        backup_get_string(mqtt, "tlsCertfile",
-                          config->mqtt.tls_certfile,
-                          sizeof(config->mqtt.tls_certfile)) &&
-        backup_get_string(mqtt, "tlsKeyfile",
-                          config->mqtt.tls_keyfile,
-                          sizeof(config->mqtt.tls_keyfile)) &&
-        backup_get_bool(mqtt, "commandEnabled",
-                        &config->mqtt.command_enabled) &&
-        backup_get_string(mqtt, "commandToken",
-                          config->mqtt.command_token,
-                          sizeof(config->mqtt.command_token));
-    if (valid && (config->mqtt.enabled || config->mqtt.server[0] != '\0')) {
-        valid = validateServerAddress(config->mqtt.server,
-                                      sizeof(config->mqtt.server) - 1);
-    }
-    if (valid && config->mqtt.command_token[0] != '\0') {
-        valid = validateMqttCommandToken(config->mqtt.command_token);
-    }
-
-    valid = valid &&
-        backup_get_bool(prometheus, "enabled",
-                        &config->prometheus.enabled) &&
-        backup_get_uint(prometheus, "port", UINT16_MAX, &number);
-    if (valid) config->prometheus.port = static_cast<uint16_t>(number);
-    valid = valid && validatePort(config->prometheus.port) &&
-        backup_get_string(prometheus, "allowedHosts",
-                          config->prometheus.allowed_hosts,
-                          sizeof(config->prometheus.allowed_hosts));
-
-    valid = valid &&
-        backup_get_bool(syslog, "enabled", &config->syslog.enabled) &&
-        backup_get_string(syslog, "server", config->syslog.server,
-                          sizeof(config->syslog.server)) &&
-        backup_get_uint(syslog, "port", UINT16_MAX, &number);
-    if (valid) config->syslog.port = static_cast<uint16_t>(number);
-    valid = valid && validatePort(config->syslog.port) &&
-        backup_get_uint(syslog, "transport", 2, &number);
-    if (valid) config->syslog.transport = static_cast<uint8_t>(number);
-    valid = valid &&
-        backup_get_uint(syslog, "minSeverity", 7, &number);
-    if (valid) config->syslog.min_severity = static_cast<uint8_t>(number);
-    valid = valid &&
-        backup_get_string(syslog, "hostname", config->syslog.hostname,
-                          sizeof(config->syslog.hostname));
-    if (valid && (config->syslog.enabled || config->syslog.server[0] != '\0')) {
-        valid = validateServerAddress(config->syslog.server,
-                                      sizeof(config->syslog.server) - 1);
-    }
-
-    valid = valid &&
-        backup_get_bool(notify, "enabled", &config->notify.enabled) &&
-        backup_get_uint(notify, "channels", 7, &number);
-    if (valid) config->notify.channels = static_cast<uint8_t>(number);
-    valid = valid &&
-        backup_get_string(notify, "webhookUrl",
-                          config->notify.webhook_url,
-                          sizeof(config->notify.webhook_url)) &&
-        backup_get_string(notify, "webhookSecret",
-                          config->notify.webhook_secret,
-                          sizeof(config->notify.webhook_secret)) &&
-        backup_get_string(notify, "telegramToken",
-                          config->notify.telegram_token,
-                          sizeof(config->notify.telegram_token)) &&
-        backup_get_string(notify, "telegramChatId",
-                          config->notify.telegram_chatid,
-                          sizeof(config->notify.telegram_chatid)) &&
-        backup_get_string(notify, "smtpServer",
-                          config->notify.smtp_server,
-                          sizeof(config->notify.smtp_server)) &&
-        backup_get_uint(notify, "smtpPort", UINT16_MAX, &number);
-    if (valid) config->notify.smtp_port = static_cast<uint16_t>(number);
-    valid = valid && validatePort(config->notify.smtp_port) &&
-        backup_get_uint(notify, "smtpTls", 2, &number);
-    if (valid) config->notify.smtp_tls = static_cast<uint8_t>(number);
-    valid = valid &&
-        backup_get_string(notify, "smtpUser", config->notify.smtp_user,
-                          sizeof(config->notify.smtp_user)) &&
-        backup_get_string(notify, "smtpPassword",
-                          config->notify.smtp_password,
-                          sizeof(config->notify.smtp_password)) &&
-        backup_get_string(notify, "smtpFrom", config->notify.smtp_from,
-                          sizeof(config->notify.smtp_from)) &&
-        backup_get_string(notify, "smtpTo", config->notify.smtp_to,
-                          sizeof(config->notify.smtp_to)) &&
-        backup_get_uint(notify, "cooldownSeconds", UINT16_MAX, &number);
-    if (valid) {
-        config->notify.cooldown_seconds = static_cast<uint16_t>(number);
-    }
-    // Optional on purpose: backups written before the event selection existed
-    // have no such key, and requiring it would make every one of them fail to
-    // restore the whole notify block. A missing key keeps the default, which
-    // is the same "everything" behaviour those backups were taken under.
-    unsigned event_mask = 0;
-    if (backup_get_uint(notify, "eventMask", NOTIFY_EVENT_ALL, &event_mask)) {
-        config->notify.event_mask = static_cast<uint16_t>(event_mask);
-    }
-    // Same CRLF/control-char rejection as the normal POST /api/monitoring
-    // path (monitoring_api.cpp): these three fields are interpolated
-    // verbatim into raw SMTP protocol lines / an HTTP header, so a crafted
-    // or corrupted backup file must not be able to reintroduce header/command
-    // injection through the restore path.
-    if (valid && config->notify.webhook_secret[0] != '\0') {
-        valid = validateNoControlChars(config->notify.webhook_secret);
-    }
-    if (valid && config->notify.smtp_from[0] != '\0') {
-        valid = validateNoControlChars(config->notify.smtp_from);
-    }
-    if (valid && config->notify.smtp_to[0] != '\0') {
-        valid = validateNoControlChars(config->notify.smtp_to);
-    }
-
-    if (!valid) {
-        delete config;
-        return ESP_ERR_INVALID_ARG;
-    }
-    *restored_config = config;
-    return ESP_OK;
-}
-
-static esp_err_t add_theme_backup(cJSON *root)
-{
-    char scheme[8] = {};
-    char color[8] = {};
-    esp_err_t result = theme_api_get_config(
-        scheme, sizeof(scheme), color, sizeof(color));
-    if (result != ESP_OK) return result;
-    cJSON *theme = cJSON_AddObjectToObject(root, "theme");
-    if (!theme) return ESP_ERR_NO_MEM;
-    cJSON_AddStringToObject(theme, "colorScheme", scheme);
-    cJSON_AddStringToObject(theme, "primaryColor", color);
-    return ESP_OK;
-}
-
-static esp_err_t parse_theme_backup(const cJSON *root,
-                                    char scheme[8], char color[8],
-                                    bool *has_theme)
-{
-    if (!scheme || !color || !has_theme) return ESP_ERR_INVALID_ARG;
-    *has_theme = false;
-    const cJSON *theme = cJSON_GetObjectItemCaseSensitive(root, "theme");
-    if (!theme) return ESP_OK; // Backwards-compatible legacy backup.
-    if (!cJSON_IsObject(theme) ||
-        !backup_get_string(theme, "colorScheme", scheme, 8) ||
-        !backup_get_string(theme, "primaryColor", color, 8)) {
-        return ESP_ERR_INVALID_ARG;
-    }
-    if (strcmp(scheme, "system") != 0 &&
-        strcmp(scheme, "light") != 0 &&
-        strcmp(scheme, "dark") != 0) {
-        return ESP_ERR_INVALID_ARG;
-    }
-    if (strlen(color) != 7 || color[0] != '#') return ESP_ERR_INVALID_ARG;
-    for (size_t index = 1; index < 7; ++index) {
-        if (!isxdigit(static_cast<unsigned char>(color[index]))) {
-            return ESP_ERR_INVALID_ARG;
-        }
-    }
-    *has_theme = true;
-    return ESP_OK;
-}
-
-esp_err_t get_backup_handler_func(httpd_req_t *req)
-{
-    add_security_headers(req);
-
-    if (validate_auth(req) != ESP_OK)
-    {
-        httpd_resp_set_status(req, "401 Not authorized");
-        httpd_resp_sendstr(req, "401 Not authorized");
-        return ESP_OK;
-    }
-
-    httpd_resp_set_type(req, "application/json");
-    httpd_resp_set_hdr(
-        req, "Content-Disposition",
-        "attachment; filename=\"hb-rf-eth-ng-backup.json\"");
-    httpd_resp_set_hdr(req, "Cache-Control",
-                       "no-store, no-cache, must-revalidate, max-age=0");
-
-    cJSON *root = cJSON_CreateObject();
-    if (!root) {
-        return httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR,
-                                   "Out of memory");
-    }
-
-    cJSON_AddStringToObject(root, "_format", "hb-rf-eth-ng-backup");
-    cJSON_AddNumberToObject(root, "_version", 2);
-
-    // A complete, text-editable JSON backup necessarily contains secrets in
-    // plaintext. Keep this warning inside the file so it remains visible even
-    // when the backup is copied or opened outside the WebUI.
-    cJSON *security = cJSON_AddObjectToObject(root, "_security");
-    if (!security) {
-        cJSON_Delete(root);
-        return httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR,
-                                   "Out of memory");
-    }
-    cJSON_AddBoolToObject(security, "containsPlaintextSecrets", true);
-    cJSON_AddStringToObject(
-        security, "warning",
-        "This unencrypted backup contains passwords, tokens, certificates and "
-        "private keys in plaintext. Store it like a password and never publish it.");
-
-    // Unknown metadata keys are intentionally ignored by the restore parser.
-    // This makes the file useful as a manually edited multi-device template
-    // without weakening validation of the actual configuration fields.
-    cJSON *portability = cJSON_AddObjectToObject(root, "_portability");
-    if (!portability) {
-        cJSON_Delete(root);
-        return httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR,
-                                   "Out of memory");
-    }
-    cJSON_AddBoolToObject(portability, "editable", true);
-    cJSON_AddStringToObject(
-        portability, "warning",
-        "Before importing on another device, review hostname, static IP, "
-        "administrator password and device-specific MQTT values.");
-
-    add_settings(root);
-
-    // Merge settings object into root if add_settings creates a sub-object
-    // NOTE: add_settings creates a "settings" object inside root.
-    // If we want a flat structure or specific structure for restore, we need to match post_settings_json_handler expectations.
-    // post_settings_json_handler expects a flat JSON object with keys like "adminUsername", "adminPassword", "hostname", etc.
-    // But add_settings creates { "settings": { "hostname": ... } }
-
-    // We need to flatten it.
-    cJSON *settingsObj = cJSON_GetObjectItem(root, "settings");
-    if (settingsObj) {
-        cJSON *child = settingsObj->child;
-        while (child) {
-            cJSON_AddItemToObject(root, child->string, cJSON_Duplicate(child, 1));
-            child = child->next;
-        }
-        cJSON_DeleteItemFromObject(root, "settings");
-    }
-
-    // A complete backup intentionally contains credentials and private keys.
-    // It is available only through the authenticated endpoint and marked
-    // no-store, but the downloaded file itself must be protected by the user.
-    cJSON_AddStringToObject(root, "adminPassword",
-                            _settings->getAdminPassword());
-    cJSON_AddBoolToObject(root, "passwordChanged",
-                          _settings->getPasswordChanged());
-
-    if (add_theme_backup(root) != ESP_OK ||
-        add_monitoring_backup(root) != ESP_OK) {
-        cJSON_Delete(root);
-        return httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR,
-                                   "Could not create complete backup");
-    }
-
-    const char *json = cJSON_PrintUnformatted(root);
-    if (json) {
-        httpd_resp_sendstr(req, json);
-        free((void *)json);
-    } else {
-        httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "JSON alloc failed");
-    }
-    cJSON_Delete(root);
-
-    return ESP_OK;
-}
-
-httpd_uri_t get_backup_handler = {
-    .uri = "/api/backup",
-    .method = HTTP_GET,
-    .handler = get_backup_handler_func,
-    .user_ctx = NULL};
-
-// Serializes restore/factory-reset mutations with monitoring configuration,
-// manual firmware writes and MQTT restarts.  Every early return releases the
-// reservation automatically; a successful reserved restart never returns.
-class ScopedOperationReservation {
-public:
-    ScopedOperationReservation() : held(ota_operation_try_begin()) {}
-    ~ScopedOperationReservation()
-    {
-        if (held) ota_operation_finish();
-    }
-    explicit operator bool() const { return held; }
-
-private:
-    bool held;
-};
-
-esp_err_t post_restore_handler_func(httpd_req_t *req)
-{
-    add_security_headers(req);
-
-    if (validate_auth(req) != ESP_OK)
-    {
-        httpd_resp_set_status(req, "401 Not authorized");
-        httpd_resp_sendstr(req, "401 Not authorized");
-        return ESP_OK;
-    }
-
-    ScopedOperationReservation operation;
-    if (!operation) {
-        httpd_resp_set_status(req, "503 Service Unavailable");
-        return httpd_resp_sendstr(
-            req, "Another configuration or restart operation is active");
-    }
-
-    // Complete backups include MQTT TLS material and notification credentials,
-    // so they can legitimately be larger than the old settings-only 8 KiB
-    // payload.
-    constexpr size_t maxRestoreBodySize = 24576;
-    if (req->content_len == 0)
-    {
-        return httpd_resp_send_err(req, HTTPD_400_BAD_REQUEST, "No data received");
-    }
-    if (req->content_len >= maxRestoreBodySize)
-    {
-        return httpd_resp_send_err(req, HTTPD_413_CONTENT_TOO_LARGE, "Request body too large");
-    }
-
-    const size_t bufferSize = req->content_len + 1;
-    char *buffer = (char*)malloc(bufferSize);
-    if (!buffer) {
-        return httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Memory allocation failed");
-    }
-
-    int len = recv_full_body(req, buffer, bufferSize);
-
-    if (len < 0)
-    {
-        free(buffer);
-        return httpd_resp_send_err(req, HTTPD_400_BAD_REQUEST, "Failed to receive request body");
-    }
-
-    cJSON *root = cJSON_Parse(buffer);
-    free(buffer);
-
-    if (!root) {
-         return httpd_resp_send_err(req, HTTPD_400_BAD_REQUEST, "Invalid JSON");
-    }
-
-    monitoring_config_t *parsed_monitoring = NULL;
-    const esp_err_t monitoring_parse_result =
-        parse_monitoring_backup(root, &parsed_monitoring);
-    std::unique_ptr<monitoring_config_t> restored_monitoring(
-        parsed_monitoring);
-    char restored_theme_scheme[8] = {};
-    char restored_theme_color[8] = {};
-    bool has_restored_theme = false;
-    const esp_err_t theme_parse_result =
-        parse_theme_backup(root, restored_theme_scheme,
-                           restored_theme_color, &has_restored_theme);
-    if (monitoring_parse_result != ESP_OK ||
-        theme_parse_result != ESP_OK) {
-        cJSON_Delete(root);
-        return send_json_error(req, "400 Bad Request",
-                               "invalid_backup", "backup");
-    }
-
-    char *adminUsername = cJSON_GetStringValue(cJSON_GetObjectItem(root, "adminUsername"));
-    char *adminPassword = cJSON_GetStringValue(cJSON_GetObjectItem(root, "adminPassword"));
-    char *currentPassword = cJSON_GetStringValue(
-        cJSON_GetObjectItem(root, "currentPassword"));
-    cJSON *passwordChangedItem =
-        cJSON_GetObjectItem(root, "passwordChanged");
-
-    char *hostname = cJSON_GetStringValue(cJSON_GetObjectItem(root, "hostname"));
-    // Presence guard mirrors the IPv6 fields below: without it,
-    // cJSON_GetBoolValue silently returns false on a missing key, and a
-    // partial payload reaching setNetworkSettings() would turn DHCP off.
-    // buildSettingsPayload() always sends useDHCP so this rarely matters,
-    // but the guard makes the contract explicit and future-proof.
-    bool useDHCP = cJSON_GetObjectItem(root, "useDHCP")
-                       ? cJSON_GetBoolValue(cJSON_GetObjectItem(root, "useDHCP"))
-                       : _settings->getUseDHCP();
-    ip4_addr_t localIP = cJSON_GetIPAddrValue(cJSON_GetObjectItem(root, "localIP"));
-    ip4_addr_t netmask = cJSON_GetIPAddrValue(cJSON_GetObjectItem(root, "netmask"));
-    ip4_addr_t gateway = cJSON_GetIPAddrValue(cJSON_GetObjectItem(root, "gateway"));
-    ip4_addr_t dns1 = cJSON_GetIPAddrValue(cJSON_GetObjectItem(root, "dns1"));
-    ip4_addr_t dns2 = cJSON_GetIPAddrValue(cJSON_GetObjectItem(root, "dns2"));
-
-    timesource_t timesource = (timesource_t)cJSON_GetIntValueSafe(
-        cJSON_GetObjectItem(root, "timesource"), (int)_settings->getTimesource());
-
-    int dcfOffset = cJSON_GetIntValueSafe(
-        cJSON_GetObjectItem(root, "dcfOffset"), _settings->getDcfOffset());
-
-    int gpsBaudrate = cJSON_GetIntValueSafe(
-        cJSON_GetObjectItem(root, "gpsBaudrate"), _settings->getGpsBaudrate());
-
-    char *ntpServer = cJSON_GetStringValue(cJSON_GetObjectItem(root, "ntpServer"));
-    int ledBrightness = cJSON_GetIntValueSafe(
-        cJSON_GetObjectItem(root, "ledBrightness"), _settings->getLEDBrightness());
-    cJSON *ledPrograms = cJSON_GetObjectItem(root, "ledPrograms");
-    bool enableIPv6 = cJSON_GetBoolValue(
-        cJSON_GetObjectItem(root, "enableIPv6"));
-    char *ipv6Mode = cJSON_GetStringValue(
-        cJSON_GetObjectItem(root, "ipv6Mode"));
-    char *ipv6Address = cJSON_GetStringValue(
-        cJSON_GetObjectItem(root, "ipv6Address"));
-    int ipv6PrefixLength = cJSON_GetIntValueSafe(
-        cJSON_GetObjectItem(root, "ipv6PrefixLength"), 64);
-    char *ipv6Gateway = cJSON_GetStringValue(cJSON_GetObjectItem(root, "ipv6Gateway"));
-    char *ipv6Dns1 = cJSON_GetStringValue(cJSON_GetObjectItem(root, "ipv6Dns1"));
-    char *ipv6Dns2 = cJSON_GetStringValue(cJSON_GetObjectItem(root, "ipv6Dns2"));
-    char *ccuIP = cJSON_GetStringValue(cJSON_GetObjectItem(root, "ccuIP"));
-    cJSON *systemLogEnabledItem =
-        cJSON_GetObjectItem(root, "systemLogEnabled");
-    cJSON *flashPauseItem = cJSON_GetObjectItem(root, "flashPause");
-    cJSON *testDesignItem = cJSON_GetObjectItem(root, "testDesignEnabled");
-
-    // Validate the entire backup and prove NVS capacity before changing any
-    // live or persistent setting. A rejected restore must leave the running
-    // configuration untouched.
-    const bool has_password_changed_flag =
-        passwordChangedItem && cJSON_IsBool(passwordChangedItem);
-    if (adminUsername && adminUsername[0] != '\0' &&
-        !valid_admin_username(adminUsername)) {
-        cJSON_Delete(root);
-        return send_json_error(req, "400 Bad Request", "invalid_username",
-                               "adminUsername");
-    }
-    bool admin_password_change_requested = false;
-    if (adminPassword && adminPassword[0] != '\0') {
-        const size_t password_len = strlen(adminPassword);
-        if (password_len >= 33) {
-            cJSON_Delete(root);
-            return send_json_error(req, "400 Bad Request", "invalid_password",
-                                   "adminPassword");
-        }
-        admin_password_change_requested =
-            secure_strcmp(adminPassword,
-                          _settings->getAdminPassword()) != 0;
-        // A same-password backup is configuration-only. For an actual
-        // credential replacement, legacy backups with an explicit flag may
-        // retain their historic password while new-format input must satisfy
-        // the current password policy.
-        const bool password_valid = !admin_password_change_requested ||
-            (has_password_changed_flag
-                ? password_len > 0
-                : validateAdminPassword(adminPassword));
-        if (!password_valid) {
-            cJSON_Delete(root);
-            return send_json_error(req, "400 Bad Request", "invalid_password",
-                                   "adminPassword");
-        }
-    }
-    if (admin_password_change_requested) {
-        if (currentPassword == NULL || currentPassword[0] == '\0') {
-            cJSON_Delete(root);
-            return send_json_error(req, "400 Bad Request",
-                                   "current_password_required",
-                                   "currentPassword");
-        }
-        if (secure_strcmp(currentPassword,
-                          _settings->getAdminPassword()) != 0) {
-            cJSON_Delete(root);
-            return send_json_error(req, "403 Forbidden",
-                                   "current_password_incorrect",
-                                   "currentPassword");
-        }
-    }
-    if (hostname && !validateHostname(hostname)) {
-        cJSON_Delete(root);
-        return send_json_error(req, "400 Bad Request", "invalid_hostname",
-                               "hostname");
-    }
-    if (hostname && !useDHCP) {
-        const bool local_ok = parse_ipv4_json(
-            cJSON_GetObjectItem(root, "localIP"), false, &localIP);
-        const bool netmask_ok = parse_ipv4_json(
-            cJSON_GetObjectItem(root, "netmask"), false, &netmask);
-        const bool gateway_ok = parse_ipv4_json(
-            cJSON_GetObjectItem(root, "gateway"), false, &gateway);
-        if (!local_ok || localIP.addr == IPADDR_ANY ||
-            !validateIPAddress(localIP) ||
-            !netmask_ok || netmask.addr == IPADDR_ANY ||
-            !validateNetmask(netmask) ||
-            !gateway_ok || !validateIPAddress(gateway)) {
-            cJSON_Delete(root);
-            return send_json_error(req, "400 Bad Request", "invalid_network",
-                                   "network");
-        }
-    }
-    if (hostname) {
-        const bool dns1_ok = parse_ipv4_json(
-            cJSON_GetObjectItem(root, "dns1"), true, &dns1);
-        const bool dns2_ok = parse_ipv4_json(
-            cJSON_GetObjectItem(root, "dns2"), true, &dns2);
-        if (!dns1_ok || !dns2_ok ||
-            (dns1.addr != IPADDR_ANY && !validateIPAddress(dns1)) ||
-            (dns2.addr != IPADDR_ANY && !validateIPAddress(dns2))) {
-            cJSON_Delete(root);
-            return send_json_error(req, "400 Bad Request", "invalid_network",
-                                   "dns");
-        }
-    }
-    if (timesource < TIMESOURCE_NTP || timesource > TIMESOURCE_GPS ||
-        !validateDcfOffset(dcfOffset) ||
-        !validateGpsBaudrate(gpsBaudrate) ||
-        !validateLEDBrightness(ledBrightness) ||
-        (ntpServer && !validateNtpServer(ntpServer)) ||
-        (ccuIP && ccuIP[0] != '\0' && !validateCcuAddress(ccuIP))) {
-        cJSON_Delete(root);
-        return send_json_error(req, "400 Bad Request", "invalid_backup",
-                               "settings");
-    }
-
-    const bool has_ipv6 =
-        cJSON_GetObjectItem(root, "enableIPv6") ||
-        cJSON_GetObjectItem(root, "ipv6Mode") ||
-        cJSON_GetObjectItem(root, "ipv6Address") ||
-        cJSON_GetObjectItem(root, "ipv6PrefixLength") ||
-        cJSON_GetObjectItem(root, "ipv6Gateway") ||
-        cJSON_GetObjectItem(root, "ipv6Dns1") ||
-        cJSON_GetObjectItem(root, "ipv6Dns2");
-    if (has_ipv6) {
-        const bool effective_enabled = cJSON_GetObjectItem(root, "enableIPv6")
-            ? enableIPv6 : _settings->getEnableIPv6();
-        const char *effective_mode =
-            ipv6Mode ? ipv6Mode : _settings->getIPv6Mode();
-        const char *effective_address =
-            ipv6Address ? ipv6Address : _settings->getIPv6Address();
-        if (!effective_mode ||
-            (strcmp(effective_mode, "auto") != 0 &&
-             strcmp(effective_mode, "static") != 0 &&
-             strcmp(effective_mode, "disabled") != 0) ||
-            ipv6PrefixLength < 1 || ipv6PrefixLength > 128 ||
-            (effective_enabled && strcmp(effective_mode, "static") == 0 &&
-             (!effective_address || effective_address[0] == '\0' ||
-              !validateIPv6Address(effective_address)))) {
-            cJSON_Delete(root);
-            return send_json_error(req, "400 Bad Request", "invalid_ipv6",
-                                   "ipv6");
-        }
-        const char *optional_ipv6[] = {
-            ipv6Gateway, ipv6Dns1, ipv6Dns2
-        };
-        for (const char *value : optional_ipv6) {
-            if (value && value[0] != '\0' && !validateIPv6Address(value)) {
-                cJSON_Delete(root);
-                return send_json_error(req, "400 Bad Request", "invalid_ipv6",
-                                       "ipv6");
-            }
-        }
-    }
-
-    static const char *LED_PROGRAM_KEYS[] = {
-        "idle", "ccu_disconnected", "ccu_connected", "update_available",
-        "error", "booting", "update_in_progress"
-    };
-    if (ledPrograms && !cJSON_IsObject(ledPrograms)) {
-        cJSON_Delete(root);
-        return send_json_error(req, "400 Bad Request", "invalid_backup",
-                               "ledPrograms");
-    }
-    if (ledPrograms) {
-        for (const char *key : LED_PROGRAM_KEYS) {
-            cJSON *item = cJSON_GetObjectItem(ledPrograms, key);
-            if (item && (!cJSON_IsNumber(item) ||
-                         item->valuedouble != item->valueint ||
-                         item->valueint < 0 || item->valueint > 10)) {
-                cJSON_Delete(root);
-                return send_json_error(req, "400 Bad Request",
-                                       "invalid_backup", "ledPrograms");
-            }
-        }
-    }
-
-    std::unique_ptr<settings_snapshot_t> previous_settings(
-        new (std::nothrow) settings_snapshot_t{});
-    std::unique_ptr<monitoring_config_t> previous_monitoring;
-    if (!previous_settings) {
-        cJSON_Delete(root);
-        return send_json_error(req, "500 Internal Server Error",
-                               "restore_snapshot_allocation", "settings");
-    }
-    _settings->snapshot(previous_settings.get());
-    if (restored_monitoring) {
-        previous_monitoring.reset(
-            new (std::nothrow) monitoring_config_t{});
-        if (!previous_monitoring ||
-            monitoring_get_config(previous_monitoring.get()) != ESP_OK) {
-            cJSON_Delete(root);
-            return send_json_error(req, "500 Internal Server Error",
-                                   "restore_snapshot_allocation",
-                                   "monitoring");
-        }
-    }
-
-    // Keep every app-owned writer out of the shared 16 KiB NVS partition
-    // from capacity preflight through the final commit. Nested helpers take
-    // this recursive lock as well.
-    NvsStorageLock restore_storage(portMAX_DELAY, "webui.restore");
-    if (!restore_storage) {
-        cJSON_Delete(root);
-        return send_json_error(req, "503 Service Unavailable",
-                               "restore_storage_busy", "backup");
-    }
-
-    if (restored_monitoring &&
-        monitoring_validate_config_storage(restored_monitoring.get()) !=
-            ESP_OK) {
-        cJSON_Delete(root);
-        return send_json_error(req, "507 Insufficient Storage",
-                               "restore_monitoring_capacity", "monitoring");
-    }
-
-    // Snapshot the rollbackable theme before changing any persistent value.
-    char previous_theme_scheme[8] = {};
-    char previous_theme_color[8] = {};
-    if (has_restored_theme) {
-        if (theme_api_get_config(previous_theme_scheme,
-                                 sizeof(previous_theme_scheme),
-                                 previous_theme_color,
-                                 sizeof(previous_theme_color)) != ESP_OK) {
-            cJSON_Delete(root);
-            return send_json_error(req, "500 Internal Server Error",
-                                   "restore_theme_snapshot_failed", "theme");
-        }
-    }
-
-    // Rotate before accepting the replacement password in RAM and before the
-    // restore transaction captures its auth rollback image. The backup will
-    // therefore contain the old durable password paired with the new token.
-    if (admin_password_change_requested) {
-        const esp_err_t token_result = rotate_admin_token();
-        if (token_result != ESP_OK) {
-            ESP_LOGE(TAG, "Restore password token rotation failed: %s",
-                     esp_err_to_name(token_result));
-            cJSON_Delete(root);
-            return send_json_error(req, "500 Internal Server Error",
-                                   "restore_token_rotation_failed",
-                                   "adminPassword");
-        }
-    }
-
-    // Apply the already-validated Settings candidate in RAM, then require a
-    // successful NVS commit before touching the other namespaces.
-    if (adminUsername && adminUsername[0] != '\0') {
-        (void)_settings->setAdminUsername(adminUsername);
-    }
-    if (admin_password_change_requested) {
-        if (has_password_changed_flag) {
-            (void)_settings->restoreAdminPassword(
-                adminPassword, cJSON_IsTrue(passwordChangedItem));
-        } else {
-            (void)_settings->setAdminPassword(adminPassword);
-        }
-    }
-    if (hostname) {
-        (void)_settings->setNetworkSettings(
-            hostname, useDHCP, localIP, netmask, gateway, dns1, dns2);
-    }
-    _settings->setTimesource(timesource);
-    _settings->setDcfOffset(dcfOffset);
-    _settings->setGpsBaudrate(gpsBaudrate);
-    if (ntpServer) _settings->setNtpServer(ntpServer);
-    _settings->setLEDBrightness(ledBrightness);
-
-    if (ledPrograms) {
-        for (int index = 0; index < 7; ++index) {
-            cJSON *item =
-                cJSON_GetObjectItem(ledPrograms, LED_PROGRAM_KEYS[index]);
-            if (item) {
-                _settings->setLedProgram(index, item->valueint);
-            }
-        }
-    }
-
-    if (has_ipv6) {
-        _settings->setIPv6Settings(
-            cJSON_GetObjectItem(root, "enableIPv6")
-                ? enableIPv6 : _settings->getEnableIPv6(),
-            ipv6Mode ? ipv6Mode : _settings->getIPv6Mode(),
-            ipv6Address ? ipv6Address : _settings->getIPv6Address(),
-            cJSON_GetObjectItem(root, "ipv6PrefixLength")
-                ? ipv6PrefixLength : _settings->getIPv6PrefixLength(),
-            ipv6Gateway ? ipv6Gateway : _settings->getIPv6Gateway(),
-            ipv6Dns1 ? ipv6Dns1 : _settings->getIPv6Dns1(),
-            ipv6Dns2 ? ipv6Dns2 : _settings->getIPv6Dns2());
-    }
-    if (ccuIP) _settings->setCCUIP(ccuIP);
-    if (systemLogEnabledItem && cJSON_IsBool(systemLogEnabledItem)) {
-        _settings->setSystemLogEnabled(cJSON_IsTrue(systemLogEnabledItem));
-    }
-    if (flashPauseItem && cJSON_IsBool(flashPauseItem)) {
-        _settings->setFlashPause(cJSON_IsTrue(flashPauseItem));
-    }
-    if (testDesignItem && cJSON_IsBool(testDesignItem)) {
-        _settings->setTestDesignEnabled(cJSON_IsTrue(testDesignItem));
-    }
-
-    const esp_err_t settings_capacity_result =
-        _settings->validateStorageCapacity();
-    if (settings_capacity_result != ESP_OK) {
-        _settings->restoreSnapshot(previous_settings.get());
-        cJSON_Delete(root);
-        if (settings_capacity_result == ESP_ERR_NVS_NOT_ENOUGH_SPACE) {
-            return send_json_error(req, "507 Insufficient Storage",
-                                   "restore_settings_capacity", "settings");
-        }
-        return send_json_error(req, "500 Internal Server Error",
-                               "restore_settings_validation_failed",
-                               "settings");
-    }
-
-    // Arm one durable marker before the first persistent Settings write. It
-    // covers Settings, theme and monitoring as a single restore generation.
-    // restore_storage remains held until the marker is cleared after all
-    // writes, or after a fully successful rollback.
-    const esp_err_t restore_begin_result =
-        Settings::beginRestoreTransaction();
-    if (restore_begin_result != ESP_OK) {
-        _settings->restoreSnapshot(previous_settings.get());
-        cJSON_Delete(root);
-        if (restore_begin_result == ESP_ERR_NVS_NOT_ENOUGH_SPACE) {
-            return send_json_error(req, "507 Insufficient Storage",
-                                   "restore_storage_capacity", "backup");
-        }
-        return send_json_error(req, "500 Internal Server Error",
-                               "restore_transaction_begin_failed", "backup");
-    }
-
-    auto finish_rollback_transaction = [](
-        const char *context, esp_err_t settings_rollback,
-        esp_err_t monitoring_rollback, esp_err_t theme_rollback) -> bool {
-        if (settings_rollback != ESP_OK || monitoring_rollback != ESP_OK ||
-            theme_rollback != ESP_OK) {
-            ESP_LOGE(TAG,
-                     "%s; restore marker retained: settings=%s monitoring=%s theme=%s",
-                     context, esp_err_to_name(settings_rollback),
-                     esp_err_to_name(monitoring_rollback),
-                     esp_err_to_name(theme_rollback));
-            return false;
-        }
-        const esp_err_t finish_result =
-            Settings::finishRestoreTransaction();
-        if (finish_result != ESP_OK) {
-            ESP_LOGE(TAG, "%s; could not clear restore marker: %s",
-                     context, esp_err_to_name(finish_result));
-            return false;
-        }
-        return true;
-    };
-
-    const esp_err_t settings_save_result = _settings->save();
-    if (settings_save_result != ESP_OK) {
-        _settings->restoreSnapshot(previous_settings.get());
-        const esp_err_t rollback_result = _settings->save();
-        const bool rollback_complete = finish_rollback_transaction(
-            "Could not roll back Settings after failed restore",
-            rollback_result, ESP_OK, ESP_OK);
-        cJSON_Delete(root);
-        if (!rollback_complete) {
-            return send_json_error(req, "500 Internal Server Error",
-                                   "restore_rollback_failed", "backup");
-        }
-        if (settings_save_result == ESP_ERR_NVS_NOT_ENOUGH_SPACE) {
-            return send_json_error(req, "507 Insufficient Storage",
-                                   "restore_settings_failed", "settings");
-        }
-        return send_json_error(req, "500 Internal Server Error",
-                               "restore_settings_failed", "settings");
-    }
-
-    bool theme_written = false;
-    if (has_restored_theme) {
-        const esp_err_t theme_result = theme_api_set_config(
-            restored_theme_scheme, restored_theme_color);
-        if (theme_result != ESP_OK) {
-            const esp_err_t theme_rollback = theme_api_set_config(
-                previous_theme_scheme, previous_theme_color);
-            _settings->restoreSnapshot(previous_settings.get());
-            const esp_err_t rollback_result = _settings->save();
-            const bool rollback_complete = finish_rollback_transaction(
-                "Could not fully roll back theme restore failure",
-                rollback_result, ESP_OK, theme_rollback);
-            cJSON_Delete(root);
-            if (!rollback_complete) {
-                return send_json_error(req, "500 Internal Server Error",
-                                       "restore_rollback_failed", "backup");
-            }
-            if (theme_result == ESP_ERR_NVS_NOT_ENOUGH_SPACE) {
-                return send_json_error(req, "507 Insufficient Storage",
-                                       "restore_theme_failed", "theme");
-            }
-            return send_json_error(req, "500 Internal Server Error",
-                                   "restore_theme_failed", "theme");
-        }
-        theme_written = true;
-    }
-
-    if (restored_monitoring) {
-        // Settings and theme share the same physical NVS partition with
-        // monitoring. Repeat the earlier preflight immediately before the
-        // monitoring transaction, while restore_storage still excludes every
-        // other app-owned writer. This catches capacity consumed by the two
-        // preceding commits without ever erasing the known-good monitoring
-        // generation.
-        const esp_err_t final_monitoring_capacity =
-            monitoring_validate_config_storage(restored_monitoring.get());
-        if (final_monitoring_capacity != ESP_OK) {
-            _settings->restoreSnapshot(previous_settings.get());
-            const esp_err_t settings_rollback = _settings->save();
-            esp_err_t theme_rollback = ESP_OK;
-            if (theme_written) {
-                theme_rollback = theme_api_set_config(
-                    previous_theme_scheme, previous_theme_color);
-            }
-            const bool rollback_complete = finish_rollback_transaction(
-                "Restore capacity rollback incomplete", settings_rollback,
-                ESP_OK, theme_rollback);
-            cJSON_Delete(root);
-            if (!rollback_complete) {
-                return send_json_error(req, "500 Internal Server Error",
-                                       "restore_rollback_failed", "backup");
-            }
-
-            if (final_monitoring_capacity ==
-                ESP_ERR_NVS_NOT_ENOUGH_SPACE) {
-                return send_json_error(req, "507 Insufficient Storage",
-                                       "restore_monitoring_capacity",
-                                       "monitoring");
-            }
-            ESP_LOGE(TAG, "Final monitoring capacity validation failed: %s",
-                     esp_err_to_name(final_monitoring_capacity));
-            return send_json_error(req, "500 Internal Server Error",
-                                   "restore_monitoring_validation_failed",
-                                   "monitoring");
-        }
-
-        const esp_err_t monitoring_result =
-            monitoring_save_config_for_restore(restored_monitoring.get());
-        if (monitoring_result != ESP_OK) {
-            _settings->restoreSnapshot(previous_settings.get());
-            const esp_err_t settings_rollback = _settings->save();
-            const esp_err_t monitoring_rollback =
-                monitoring_save_config_for_restore(previous_monitoring.get());
-            esp_err_t theme_rollback = ESP_OK;
-            if (theme_written) {
-                theme_rollback = theme_api_set_config(
-                    previous_theme_scheme, previous_theme_color);
-            }
-            const bool rollback_complete = finish_rollback_transaction(
-                "Restore rollback incomplete", settings_rollback,
-                monitoring_rollback, theme_rollback);
-            cJSON_Delete(root);
-            if (!rollback_complete) {
-                return send_json_error(req, "500 Internal Server Error",
-                                       "restore_rollback_failed", "backup");
-            }
-            if (monitoring_result == ESP_ERR_NVS_NOT_ENOUGH_SPACE) {
-                return send_json_error(req, "507 Insufficient Storage",
-                                       "restore_monitoring_capacity",
-                                       "monitoring");
-            }
-            return send_json_error(req, "500 Internal Server Error",
-                                   "restore_monitoring_failed", "monitoring");
-        }
-    }
-
-    const esp_err_t restore_finish_result =
-        Settings::finishRestoreTransaction();
-    if (restore_finish_result != ESP_OK) {
-        _settings->restoreSnapshot(previous_settings.get());
-        const esp_err_t settings_rollback = _settings->save();
-        esp_err_t monitoring_rollback = ESP_OK;
-        if (restored_monitoring) {
-            monitoring_rollback =
-                monitoring_save_config_for_restore(previous_monitoring.get());
-        }
-        esp_err_t theme_rollback = ESP_OK;
-        if (theme_written) {
-            theme_rollback = theme_api_set_config(
-                previous_theme_scheme, previous_theme_color);
-        }
-        const bool rollback_complete = finish_rollback_transaction(
-            "Restore marker finalization failed", settings_rollback,
-            monitoring_rollback, theme_rollback);
-        cJSON_Delete(root);
-        if (!rollback_complete) {
-            return send_json_error(req, "500 Internal Server Error",
-                                   "restore_rollback_failed", "backup");
-        }
-        return send_json_error(req, "500 Internal Server Error",
-                               "restore_transaction_finish_failed", "backup");
-    }
-
-    // Only expose non-persistent side effects after every namespace committed.
-    if (cJSON_GetObjectItem(root, "ledBrightness") != NULL) {
-        LED::setBrightness(ledBrightness);
-    }
-    if (flashPauseItem && cJSON_IsBool(flashPauseItem)) {
-        set_flash_pause_enabled(cJSON_IsTrue(flashPauseItem));
-    }
-
-    // Do not carry the storage lock into the restart cleanup: a late MQTT
-    // callback may need to write its one-shot migration marker before the
-    // cooperative MQTT stop can finish.
-    restore_storage.release();
-
-    cJSON_Delete(root);
-
-    httpd_resp_set_type(req, "application/json");
-    httpd_resp_sendstr(req, "{\"success\":true}");
-
-    // Restart
-    vTaskDelay(pdMS_TO_TICKS(1000));
-    refresh_restart_sync_from_settings();
-    full_system_restart_with_reserved_operation();
-
-    return ESP_OK;
-}
-
-httpd_uri_t post_restore_handler = {
-    .uri = "/api/restore",
-    .method = HTTP_POST,
-    .handler = post_restore_handler_func,
-    .user_ctx = NULL};
-
-// Forward declaration: prepare_ota_heap() is defined further down (after the
-// upload handler) but is now also used on the success path of the upload
-// handler so that heap/network-active subsystems are stopped before the
-// restart. Without this the upload handler would not compile.
-static esp_err_t prepare_ota_heap(uint32_t *paused_monitoring);
-struct ota_finalize_ctx {
-    const esp_partition_t *update_partition;
-    const esp_partition_t *running;
-};
-static void ota_finalize_task(void *pv);
-
-#define OTA_CHECK(a, str, ...)                                                    \
-    do                                                                            \
-    {                                                                             \
-        if (!(a))                                                                 \
-        {                                                                         \
-            ESP_LOGE(TAG, "%s(%d): " str, __FUNCTION__, __LINE__, ##__VA_ARGS__); \
-            httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, str);       \
-            goto err;                                                             \
-        }                                                                         \
-    } while (0)
-
-#define OTA_BUFFER_SIZE 4096
-
-// OTA status tracking for the remaining manual firmware upload endpoint
-// (/ota_update). Automatic update checks and URL-based OTA were removed, but
-// this gate still prevents two administrators from writing the update
-// partition concurrently.
-enum ota_status_t {
-    OTA_IDLE = 0,
-    OTA_DOWNLOADING,
-    OTA_FINALIZING,
-    OTA_SUCCESS,
-    OTA_FAILED
-};
-
-static std::atomic<ota_status_t> _ota_status{OTA_IDLE};
-static std::atomic<int> _ota_progress{0};  // 0-100
-static char _ota_error[128] = {0};
-static portMUX_TYPE _ota_error_mux = portMUX_INITIALIZER_UNLOCKED;
-
-static void set_ota_error(const char *format, ...)
-{
-    char text[sizeof(_ota_error)];
-    va_list args;
-    va_start(args, format);
-    vsnprintf(text, sizeof(text), format, args);
-    va_end(args);
-
-    portENTER_CRITICAL(&_ota_error_mux);
-    snprintf(_ota_error, sizeof(_ota_error), "%s", text);
-    portEXIT_CRITICAL(&_ota_error_mux);
-}
-
-static void copy_ota_error(char *dest, size_t size)
-{
-    portENTER_CRITICAL(&_ota_error_mux);
-    snprintf(dest, size, "%s", _ota_error);
-    portEXIT_CRITICAL(&_ota_error_mux);
-}
-
-esp_err_t post_ota_update_handler_func(httpd_req_t *req)
-{
-    add_security_headers(req);
-    bool net_gate_held = false;
-
-    if (validate_auth(req) != ESP_OK)
-    {
-        httpd_resp_set_status(req, "401 Not authorized");
-        httpd_resp_sendstr(req, "401 Not authorized");
-        return ESP_OK;
-    }
-
-    if (!ota_operation_try_begin())
-    {
-        httpd_resp_send_err(req, HTTPD_400_BAD_REQUEST, "OTA update already in progress");
-        return ESP_OK;
-    }
-    _ota_status = OTA_DOWNLOADING;
-    _ota_progress = 0;
-    set_ota_error("");
-
-    // The inbound upload itself is not TLS, but its 4 KiB buffer plus flash
-    // erase/write pressure must not overlap an outbound TLS handshake on a
-    // WROOM-32. Mark the operation before waiting so best-effort consumers
-    // defer; hold the ownershipless gate only while bytes are streamed.
-    net_fetch_set_ota_active(true);
-    if (g_net_fetch_mutex != NULL) {
-        if (xSemaphoreTake(g_net_fetch_mutex,
-                           pdMS_TO_TICKS(20000)) != pdTRUE) {
-            ESP_LOGE(TAG, "Could not reserve network/TLS gate for OTA upload");
-            _ota_status = OTA_FAILED;
-            set_ota_error("Network/TLS subsystem busy");
-            net_fetch_set_ota_active(false);
-            ota_operation_finish();
-            return httpd_resp_send_err(req,
-                                       HTTPD_500_INTERNAL_SERVER_ERROR,
-                                       "Network/TLS subsystem busy");
-        }
-        net_gate_held = true;
-        crash_blackbox_net_op_begin("webui_ota_upload");
-    }
-
-    esp_ota_handle_t ota_handle = 0;
-    bool ota_begun = false;
-
-    char *ota_buff = (char *)malloc(OTA_BUFFER_SIZE);
-    if (!ota_buff) {
-        ESP_LOGE(TAG, "Failed to allocate OTA buffer");
-        httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Memory allocation failed");
-        _ota_status = OTA_FAILED;
-        if (net_gate_held) { crash_blackbox_net_op_end(); xSemaphoreGive(g_net_fetch_mutex); }
-        net_fetch_set_ota_active(false);
-        ota_operation_finish();
-        return ESP_FAIL;
-    }
-
-    const size_t content_length = req->content_len;
-    if (content_length == 0x50000) {
-        ESP_LOGW(TAG, "Rejected 320 KiB WebUI image on firmware endpoint");
-        free(ota_buff);
-        _ota_status = OTA_FAILED;
-        if (net_gate_held) { crash_blackbox_net_op_end(); xSemaphoreGive(g_net_fetch_mutex); }
-        net_fetch_set_ota_active(false);
-        ota_operation_finish();
-        return httpd_resp_send_err(req, HTTPD_400_BAD_REQUEST,
-            "Falsche Datei: Das 327680-Byte-WebUI-/WWW-Image muss unter System -> WebUI installiert werden.");
-    }
-    size_t content_received = 0;
-    int recv_len;
-    int timeout_retries = 5;
-    static constexpr int64_t OTA_UPLOAD_TOTAL_TIMEOUT_US =
-        5LL * 60LL * 1000LL * 1000LL;
-    static constexpr int64_t OTA_UPLOAD_NO_PROGRESS_TIMEOUT_US =
-        30LL * 1000LL * 1000LL;
-    const int64_t upload_started_us = esp_timer_get_time();
-    int64_t last_progress_us = upload_started_us;
-    bool is_req_body_started = false;
-    const esp_partition_t *update_partition = esp_ota_get_next_update_partition(NULL);
-    const esp_partition_t *running = NULL;
-    esp_err_t ota_end_result = ESP_OK;
-
-    // Validate update partition exists
-    if (update_partition == NULL) {
-        ESP_LOGE(TAG, "No OTA update partition found");
-        httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "No OTA partition available");
-        goto err;
-    }
-
-    // HTTP Content-Length is size_t. Narrowing it to int before MIN() could
-    // turn a value above INT_MAX negative and then back into a huge size_t in
-    // httpd_req_recv(), overflowing the fixed 4 KiB buffer. Also reject images
-    // which cannot fit the selected application partition before reading a
-    // single byte from the socket.
-    if (content_length == 0 || content_length > update_partition->size ||
-        content_length > static_cast<size_t>(INT32_MAX)) {
-        ESP_LOGW(TAG, "Rejected OTA content length: %u (partition %u)",
-                 (unsigned)content_length, (unsigned)update_partition->size);
-        httpd_resp_send_err(req, HTTPD_400_BAD_REQUEST,
-                            "Invalid firmware image size");
-        goto err;
-    }
-
-    ESP_LOGI(TAG, "Starting OTA update, partition: %s, size: %u bytes",
-             update_partition->label, (unsigned)content_length);
-
-    do
-    {
-        const int64_t now_us = esp_timer_get_time();
-        if (now_us - upload_started_us > OTA_UPLOAD_TOTAL_TIMEOUT_US ||
-            now_us - last_progress_us > OTA_UPLOAD_NO_PROGRESS_TIMEOUT_US) {
-            ESP_LOGE(TAG,
-                     "OTA upload deadline exceeded after %u of %u bytes",
-                     (unsigned)content_received, (unsigned)content_length);
-            set_ota_error("Firmware upload timed out");
-            httpd_resp_set_status(req, "408 Request Timeout");
-            httpd_resp_sendstr(req, "Firmware upload timed out");
-            goto err;
-        }
-
-        const size_t receive_cap =
-            MIN(content_length - content_received,
-                static_cast<size_t>(OTA_BUFFER_SIZE));
-        if ((recv_len = httpd_req_recv(req, ota_buff, receive_cap)) < 0)
-        {
-            if (recv_len == HTTPD_SOCK_ERR_TIMEOUT && timeout_retries-- > 0)
-            {
-                // Transient timeout - retry a bounded number of times. An
-                // unbounded retry loop would wedge the single httpd task
-                // forever if the client stalls mid-upload.
-                continue;
-            }
-            else
-            {
-                ESP_LOGE(TAG, "OTA socket error %d, received %u of %u bytes",
-                         recv_len, (unsigned)content_received,
-                         (unsigned)content_length);
-                httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Network error during upload");
-                goto err;
-            }
-        }
-        else if (recv_len == 0)
-        {
-            // Connection closed by client
-            ESP_LOGE(TAG,
-                     "OTA connection closed prematurely, received %u of %u bytes",
-                     (unsigned)content_received, (unsigned)content_length);
-            httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Incomplete upload");
-            goto err;
-        }
-
-        // A peer may trickle a few bytes before each socket timeout. Reset the
-        // transient retry budget on real progress, while the independent
-        // no-progress and total deadlines keep the single httpd task bounded.
-        last_progress_us = esp_timer_get_time();
-        timeout_retries = 5;
-
-        if (!is_req_body_started)
-        {
-            is_req_body_started = true;
-
-            if (recv_len <= 0 || static_cast<unsigned char>(ota_buff[0]) != 0xE9) {
-                ESP_LOGW(TAG, "Rejected non-ESP firmware image (magic 0x%02x)",
-                         recv_len > 0 ? static_cast<unsigned char>(ota_buff[0]) : 0);
-                httpd_resp_send_err(req, HTTPD_400_BAD_REQUEST,
-                    "Falsche Datei: kein gueltiges ESP32-Firmware-Abbild. WebUI unter System -> WebUI installieren.");
-                goto err;
-            }
-
-            // Only raw binary uploads are supported (the WebUI posts the file
-            // as the request body). The previous multipart/form-data path was
-            // broken by design: it compared stripped body bytes against the
-            // full content length (loop never terminated) and wrote the
-            // trailing boundary into flash.
-            char content_type[64] = {0};
-            if (httpd_req_get_hdr_value_str(req, "Content-Type", content_type, sizeof(content_type)) == ESP_OK &&
-                strstr(content_type, "multipart/form-data") != NULL) {
-                ESP_LOGE(TAG, "Multipart firmware uploads are not supported - send the raw binary as request body");
-                httpd_resp_send_err(req, HTTPD_400_BAD_REQUEST, "Multipart uploads not supported, send raw binary body");
-                goto err;
-            }
-
-            OTA_CHECK(esp_ota_begin(update_partition, content_length, &ota_handle) == ESP_OK, "Could not start OTA");
-            ota_begun = true;
-            ESP_LOGW(TAG, "Begin OTA Update to partition %s, File Size: %u",
-                     update_partition->label, (unsigned)content_length);
-            _statusLED->setState(LED_STATE_BLINK_FAST);
-
-            OTA_CHECK(esp_ota_write(ota_handle, ota_buff, recv_len) == ESP_OK, "Error writing OTA");
-            content_received += static_cast<size_t>(recv_len);
-            _ota_progress = static_cast<int>(content_received * 100 /
-                                             content_length);
-            ESP_LOGI(TAG, "OTA progress: %u / %u bytes (%d%%)",
-                     (unsigned)content_received, (unsigned)content_length,
-                     _ota_progress.load());
-        }
-        else
-        {
-            OTA_CHECK(esp_ota_write(ota_handle, ota_buff, recv_len) == ESP_OK, "Error writing OTA");
-            content_received += static_cast<size_t>(recv_len);
-            _ota_progress = static_cast<int>(content_received * 100 /
-                                             content_length);
-            ESP_LOGI(TAG, "OTA progress: %u / %u bytes (%d%%)",
-                     (unsigned)content_received, (unsigned)content_length,
-                     _ota_progress.load());
-        }
-    } while (content_received < content_length);
-
-    // Verify complete firmware was received
-    if (content_received != content_length) {
-        ESP_LOGE(TAG, "Incomplete firmware: received %u of %u bytes",
-                 (unsigned)content_received, (unsigned)content_length);
-        httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Incomplete firmware upload");
-        goto err;
-    }
-
-    // Validate and finalize OTA
-    // esp_ota_end() consumes/removes the OTA handle even when image
-    // verification fails. Clear the ownership flag before testing its result
-    // so the error path never calls esp_ota_abort() on a stale handle.
-    ota_end_result = esp_ota_end(ota_handle);
-    ota_begun = false;
-    OTA_CHECK(ota_end_result == ESP_OK, "Error finalizing OTA");
-
-    // Verify the firmware image before setting boot partition
-    ESP_LOGI(TAG, "Validating firmware image...");
-    running = esp_ota_get_running_partition();
-
-    if (running == NULL || update_partition == running) {
-        ESP_LOGE(TAG, "Cannot update running partition!");
-        httpd_resp_send_err(req, HTTPD_500_INTERNAL_SERVER_ERROR, "Invalid OTA partition");
-        goto err;
-    }
-
-    // The image is finalized and no longer needs the upload buffer. Release it
-    // before stopping workers so a fragmented heap does not turn an otherwise
-    // valid update into a cleanup timeout.
-    free(ota_buff);
-    ota_buff = NULL;
-
-    // Release before stopping MQTT/other workers: an MQTT reconnect callback
-    // may be waiting on this gate and its cooperative stop must be allowed to
-    // finish. No more upload buffer or flash write is active beyond this point.
-    if (net_gate_held) {
-        crash_blackbox_net_op_end();
-        xSemaphoreGive(g_net_fetch_mutex);
-        net_gate_held = false;
-    }
-    net_fetch_set_ota_active(false);
-
-    // Offload the post-upload sequence (worker stop → boot select → restart)
-    // to a background task so the httpd thread stays free to serve status
-    // polls during the potentially 90 s sequential worker-stop phase.
-    _ota_status = OTA_FINALIZING;
-
-    httpd_resp_set_type(req, "application/json");
-    httpd_resp_sendstr(req, "{\"success\":true,\"message\":\"Firmware received, finalizing and restarting...\"}");
-
-    // Scope ctx so goto err (from the upload phase above) cannot jump over
-    // its initialization — C++ forbids crossing a declaration with an
-    // initializer even for pointer types.
-    {
-        ota_finalize_ctx *ctx = (ota_finalize_ctx *)calloc(1, sizeof(ota_finalize_ctx));
-        if (!ctx) {
-            ESP_LOGE(TAG, "Could not allocate OTA finalize context");
-            _ota_status = OTA_FAILED;
-            set_ota_error("Memory allocation failed during finalize");
-            ota_operation_finish();
-            return ESP_OK;
-        }
-        ctx->update_partition = update_partition;
-        ctx->running = running;
-
-        if (xTaskCreate(ota_finalize_task, "ota_finalize", 8192,
-                        ctx, 2, NULL) != pdPASS) {
-            ESP_LOGE(TAG, "Could not create OTA finalize task");
-            _ota_status = OTA_FAILED;
-            set_ota_error("Could not start finalize task");
-            free(ctx);
-            ota_operation_finish();
-            return ESP_OK;
-        }
-    }
-
-    return ESP_OK;
-
-err:
-    if (ota_buff) free(ota_buff);
-    if (net_gate_held) { crash_blackbox_net_op_end(); xSemaphoreGive(g_net_fetch_mutex); }
-    net_fetch_set_ota_active(false);
-    _statusLED->setState(LED_STATE_OFF);
-    _ota_status = OTA_FAILED;
-
-    // Abort OTA if it was started but not completed
-    if (ota_begun) {
-        ESP_LOGW(TAG, "Aborting OTA operation due to error");
-        esp_ota_abort(ota_handle);
-    }
-
-    // Store reset reason for failed firmware update
-    ResetInfo::storeResetReason(RESET_REASON_UPDATE_FAILED);
-    ota_operation_finish();
-    return ESP_FAIL;
-}
-
-httpd_uri_t post_ota_update_handler = {
-    .uri = "/ota_update",
-    .method = HTTP_POST,
-    .handler = post_ota_update_handler_func,
-    .user_ctx = NULL};
-
-esp_err_t post_restart_handler_func(httpd_req_t *req)
-{
-    add_security_headers(req);
-
-    if (validate_auth(req) != ESP_OK)
-    {
-        return httpd_resp_send_err(req, HTTPD_401_UNAUTHORIZED, NULL);
-    }
-
-    httpd_resp_set_type(req, "application/json");
-    /* CORS header removed - same-origin requests only */
-    httpd_resp_sendstr(req, "{\"success\":true}");
-
-    // Store reset reason before restart
-    ResetInfo::storeResetReason(RESET_REASON_USER_RESTART);
-
-    // Queue the notification before the delay below so the events worker has
-    // a chance to pick it up (queue poll: 500ms) and start delivery before
-    // full_system_restart()'s monitoring_pause_for_ota() stops the worker —
-    // that stop path waits up to 30s for an in-flight send to finish, but
-    // only if the worker already dequeued the entry.
-    events_emit(EVENT_RESTART, "manual restart via WebUI/API");
-
-    // Restart after a short delay to allow response to be sent
-    vTaskDelay(pdMS_TO_TICKS(1000));
-    refresh_restart_sync_from_settings();
-    full_system_restart();
-
-    return ESP_OK;
-}
-
-httpd_uri_t post_restart_handler = {
-    .uri = "/api/restart",
-    .method = HTTP_POST,
-    .handler = post_restart_handler_func,
-    .user_ctx = NULL};
-
-esp_err_t post_factory_reset_handler_func(httpd_req_t *req)
-{
-    add_security_headers(req);
-
-    if (validate_auth(req) != ESP_OK)
-    {
-        return httpd_resp_send_err(req, HTTPD_401_UNAUTHORIZED, NULL);
-    }
-
-    ScopedOperationReservation operation;
-    if (!operation) {
-        httpd_resp_set_status(req, "503 Service Unavailable");
-        return httpd_resp_sendstr(
-            req, "Another configuration or restart operation is active");
-    }
-
-    // Queue the notification against the still-live (pre-wipe) notify config
-    // before anything is erased. Same delivery-window reasoning as
-    // post_restart_handler_func: the events worker gets the 1s pre-restart
-    // delay plus up to 30s inside monitoring_pause_for_ota() to actually
-    // send it, as long as it dequeues the entry before that stop begins.
-    events_emit(EVENT_FACTORY_RESET, "factory reset via WebUI/API");
-
-    // Erase every user-controlled namespace first. reset_info is deliberately
-    // part of Settings::clear(), so store the one allowed post-reset metadata
-    // value and report success only after the complete wipe succeeded.
-    const esp_err_t clear_result = _settings->clear();
-    if (clear_result != ESP_OK) {
-        ESP_LOGE(TAG, "Factory reset failed while clearing settings: %s",
-                 esp_err_to_name(clear_result));
-        return send_json_error(req, "500 Internal Server Error",
-                               "factory_reset_failed", "settings");
-    }
-    ResetInfo::storeResetReason(RESET_REASON_FACTORY_RESET);
-
-    httpd_resp_set_type(req, "application/json");
-    /* CORS header removed - same-origin requests only */
-    httpd_resp_sendstr(req, "{\"success\":true}");
-
-    // Restart after a short delay to allow response to be sent
-    vTaskDelay(pdMS_TO_TICKS(1000));
-    refresh_restart_sync_from_settings();
-    full_system_restart_with_reserved_operation();
-
-    return ESP_OK;
-}
-
-httpd_uri_t post_factory_reset_handler = {
-    .uri = "/api/factory-reset",
-    .method = HTTP_POST,
-    .handler = post_factory_reset_handler_func,
-    .user_ctx = NULL};
-
-static esp_err_t get_ota_status_handler_func(httpd_req_t *req)
-{
-    add_security_headers(req);
-
-    if (validate_auth(req) != ESP_OK)
-    {
-        return httpd_resp_send_err(req, HTTPD_401_UNAUTHORIZED, NULL);
-    }
-
-    const char *status_str;
-    const ota_status_t status = _ota_status.load();
-    switch (status) {
-        case OTA_DOWNLOADING: status_str = "downloading"; break;
-        case OTA_FINALIZING:  status_str = "finalizing"; break;
-        case OTA_SUCCESS:     status_str = "success"; break;
-        case OTA_FAILED:      status_str = "failed"; break;
-        default:              status_str = "idle"; break;
-    }
-
-    const char* flashPause = (_settings && _settings->getFlashPause()) ? "true" : "false";
-    // The escaped OTA error can use up to twice the 128-byte source buffer.
-    // Leave enough room for the surrounding JSON without truncating it.
-    char buf[384];
-
-    if (status == OTA_FAILED) {
-        char error[sizeof(_ota_error)];
-        copy_ota_error(error, sizeof(error));
-        if (error[0] != '\0') {
-            // Escape any quotes in the error string just in case
-            char esc_error[sizeof(_ota_error) * 2] = {0};
-            int j = 0;
-            for (int i = 0; error[i] && j < sizeof(esc_error) - 2; i++) {
-                if (error[i] == '"' || error[i] == '\\') {
-                    esc_error[j++] = '\\';
-                }
-                esc_error[j++] = error[i];
-            }
-            snprintf(buf, sizeof(buf), "{\"status\":\"%s\",\"progress\":%d,\"flashPause\":%s,\"error\":\"%s\"}",
-                     status_str, _ota_progress.load(), flashPause, esc_error);
-        } else {
-            snprintf(buf, sizeof(buf), "{\"status\":\"%s\",\"progress\":%d,\"flashPause\":%s}",
-                     status_str, _ota_progress.load(), flashPause);
-        }
-    } else {
-        snprintf(buf, sizeof(buf), "{\"status\":\"%s\",\"progress\":%d,\"flashPause\":%s}",
-                 status_str, _ota_progress.load(), flashPause);
-    }
-
-    httpd_resp_set_type(req, "application/json");
-    httpd_resp_sendstr(req, buf);
-
-    return ESP_OK;
-}
-
-httpd_uri_t get_ota_status_handler = {
-    .uri = "/api/ota_status",
-    .method = HTTP_GET,
-    .handler = get_ota_status_handler_func,
-    .user_ctx = NULL};
-
-// Free heap for the manual firmware upload restart by shutting down heap-heavy
-// subsystems.
-// The ESP32-WROOM-32 has no PSRAM; with MQTT/monitoring running, only
-// ~60 KB heap can remain. Keeping those workers alive while finalizing the
-// uploaded image and taking Ethernet down also creates avoidable contention.
-// On OTA success the device restarts and
-// everything comes back; on failure the returned mask is used to resume the
-// paused monitoring workers without requiring a manual restart.
-static esp_err_t prepare_ota_heap(uint32_t *paused_monitoring)
-{
-    if (!paused_monitoring) return ESP_ERR_INVALID_ARG;
-    *paused_monitoring = 0;
-    ESP_LOGI(TAG, "Preparing heap for manual OTA restart (current free: %u KB)",
-             (unsigned)(esp_get_free_heap_size() / 1024));
-
-    // Stop MQTT, CheckMK, Prometheus, Syslog and notification workers. Besides
-    // TLS state, this can free several task stacks (6-8 KB each) before OTA.
-    esp_err_t monitoring_pause_result =
-        monitoring_pause_for_ota(paused_monitoring);
-    if (monitoring_pause_result != ESP_OK) {
-        ESP_LOGE(TAG, "Cannot continue OTA/restart while monitoring is stopping: %s",
-                 esp_err_to_name(monitoring_pause_result));
-        return monitoring_pause_result;
-    }
-
-    // Note: the automatic update-check feature (former UpdateCheck esp_timer)
-    // was removed, so there is no background fetch task left to stop here.
-
-    // Brief settle for heap de-fragmentation
-    vTaskDelay(pdMS_TO_TICKS(200));
-    return ESP_OK;
-}
-
-// Background finalize task: stop workers, select boot partition, restart.
-// Runs on its own task so the httpd thread is free to serve status polls
-// during the (potentially 90 s) sequential worker-stop phase.
-static void ota_finalize_task(void *pv)
-{
-    ota_finalize_ctx *ctx = static_cast<ota_finalize_ctx *>(pv);
-    uint32_t paused_monitoring = 0;
-
-    esp_err_t prepare_result = prepare_ota_heap(&paused_monitoring);
-    if (prepare_result != ESP_OK) {
-        _statusLED->setState(LED_STATE_OFF);
-        _ota_status = OTA_FAILED;
-        set_ota_error("Restart deferred: background network cleanup failed");
-        ResetInfo::storeResetReason(RESET_REASON_UPDATE_FAILED);
-        monitoring_resume_after_ota(paused_monitoring);
-        ota_operation_finish();
-        free(ctx);
-        vTaskDelete(NULL);
-        return;
-    }
-
-    esp_err_t boot_result = esp_ota_set_boot_partition(ctx->update_partition);
-    const esp_partition_t *selected_boot = esp_ota_get_boot_partition();
-    if (selected_boot != ctx->update_partition) {
-        ESP_LOGE(TAG, "Could not select uploaded OTA partition: set=%s selected=%s",
-                 esp_err_to_name(boot_result),
-                 selected_boot ? selected_boot->label : "unknown");
-
-        bool running_restored = (selected_boot == ctx->running);
-        for (int attempt = 0; !running_restored && attempt < 3; ++attempt) {
-            esp_err_t restore_result = esp_ota_set_boot_partition(ctx->running);
-            selected_boot = esp_ota_get_boot_partition();
-            running_restored =
-                (restore_result == ESP_OK && selected_boot == ctx->running);
-            if (!running_restored) {
-                vTaskDelay(pdMS_TO_TICKS(50));
-            }
-        }
-
-        _statusLED->setState(LED_STATE_OFF);
-        _ota_status = OTA_FAILED;
-        set_ota_error(running_restored
-                          ? "Could not activate uploaded firmware"
-                          : "Boot selection uncertain; restarting safely");
-        ResetInfo::storeResetReason(RESET_REASON_UPDATE_FAILED);
-
-        if (running_restored) {
-            monitoring_resume_after_ota(paused_monitoring);
-            ota_operation_finish();
-            free(ctx);
-            vTaskDelete(NULL);
-            return;
-        }
-
-        vTaskDelay(pdMS_TO_TICKS(1000));
-        refresh_restart_sync_from_settings();
-        full_system_restart_with_reserved_operation();
-        free(ctx);
-        return;
-    }
-
-    ESP_LOGI(TAG, "OTA finished successfully, restarting to activate new firmware.");
-    _statusLED->setState(LED_STATE_OFF);
-    _ota_progress = 100;
-    _ota_status = OTA_SUCCESS;
-    ResetInfo::storeResetReason(RESET_REASON_FIRMWARE_UPDATE);
-
-    vTaskDelay(pdMS_TO_TICKS(3000));
-    refresh_restart_sync_from_settings();
-    full_system_restart_with_reserved_operation();
-    free(ctx);
-}
 
 esp_err_t post_change_password_handler_func(httpd_req_t *req)
 {
